Chore/upgrade bevy 0.16

+++ src/entities/satellite/grabber.rs
@@ -1,5 +1,4 @@
-use std::f32::EPSILON;
-
+use bevy::math::ops::atan2;
 use bevy::prelude::*;
 use bevy::text::{JustifyText, Text2d, TextColor, TextFont, TextLayout};
 use bevy_ggrs::{GgrsSchedule, LocalPlayers};
@@ -15,7 +14,7 @@ const ROPE_MAX_LENGTH: f32 = 275.0;
 const ROPE_ADJUST_SPEED: f32 = 50.0;
 
 #[derive(Component)]
-#[require(Name(|| Name::new("Grabber")))]
+#[require(Name::new("Grabber"))]
 pub struct Grabber;
 
 #[derive(Component)]
@@ -38,13 +37,13 @@ pub struct GrabbedOrbit {
 }
 
 #[derive(Component)]
-#[require(Name(|| Name::new("PlayerPrompt")))]
+#[require(Name::new("PlayerPrompt"))]
 pub struct PlayerPrompt {
     pub player: Entity,
 }
 
 #[derive(Component)]
-#[require(Name(|| Name::new("GrabberRope")))]
+#[require(Name::new("GrabberRope"))]
 pub struct GrabberRope {
     pub player: Entity,
     pub grabber: Entity,
@@ -124,30 +123,30 @@ fn display_interact_prompt(
             continue;
         }
 
-        if prompt_query.iter().all(|p| p.player != player_entity) {
-            if let Ok(grabber_transform) = grabber_query.get(nearby_grabber.0) {
-                let font = asset_server.load("fonts/FiraSans-Bold.ttf");
-                commands.spawn((
-                    Text2d::new(prompt.message.clone()),
-                    TextFont {
-                        font,
-                        font_size: 30.0,
-                        ..Default::default()
-                    },
-                    TextColor(Color::WHITE),
-                    TextLayout {
-                        justify: JustifyText::Center,
-                        ..Default::default()
-                    },
-                    Transform::from_translation(
-                        grabber_transform.translation + Vec3::new(0.0, -100.0, 2.0),
-                    ),
-                    GlobalTransform::default(),
-                    PlayerPrompt {
-                        player: player_entity,
-                    },
-                ));
-            }
+        if prompt_query.iter().all(|p| p.player != player_entity)
+            && let Ok(grabber_transform) = grabber_query.get(nearby_grabber.0)
+        {
+            let font = asset_server.load("fonts/FiraSans-Bold.ttf");
+            commands.spawn((
+                Text2d::new(prompt.message.clone()),
+                TextFont {
+                    font,
+                    font_size: 30.0,
+                    ..Default::default()
+                },
+                TextColor(Color::WHITE),
+                TextLayout {
+                    justify: JustifyText::Center,
+                    ..Default::default()
+                },
+                Transform::from_translation(
+                    grabber_transform.translation + Vec3::new(0.0, -100.0, 2.0),
+                ),
+                GlobalTransform::default(),
+                PlayerPrompt {
+                    player: player_entity,
+                },
+            ));
         }
     }
 }
@@ -185,50 +184,50 @@ fn handle_grabber_interaction(
         let is_pressed = actions.pressed(&PlayerAction::Interact);
 
         if is_pressed && grabbed_by.is_none() {
-            if let Some(nearby) = nearby {
-                if let Ok(grabber_tf) = grabber_query.get(nearby.0) {
-                    let center = grabber_tf.translation.truncate();
-                    let pos = position.0;
-                    let offset = pos - center;
-                    let distance = offset.length();
-                    let angle = offset.y.atan2(offset.x);
+            if let Some(nearby) = nearby
+                && let Ok(grabber_tf) = grabber_query.get(nearby.0)
+            {
+                let center = grabber_tf.translation.truncate();
+                let pos = position.0;
+                let offset = pos - center;
+                let distance = offset.length();
+                let angle = atan2(offset.y, offset.x);
 
-                    let tangent = if offset.length_squared() > EPSILON {
-                        Vec2::new(-offset.y, offset.x).normalize()
-                    } else {
-                        Vec2::ZERO
-                    };
+                let tangent = if offset.length_squared() > f32::EPSILON {
+                    Vec2::new(-offset.y, offset.x).normalize()
+                } else {
+                    Vec2::ZERO
+                };
 
-                    let sign = if vel.0.dot(tangent) >= 0.0 { 1.0 } else { -1.0 };
-                    let mut speed = vel.0.dot(tangent);
-                    if speed.abs() < 600.0 {
-                        speed = 600.0 * sign;
-                    }
-
-                    commands
-                        .entity(player_entity)
-                        .insert(GrabbedBy(nearby.0))
-                        .insert(GrabbedOrbit {
-                            center,
-                            distance,
-                            angle,
-                            initial_speed: speed,
-                        });
-                    let mesh = meshes.add(Rectangle::new(4.0, 1.0));
-                    let material = materials.add(Color::srgb(0.0, 0.0, 1.0));
-
-                    commands.spawn((
-                        Mesh2d(mesh),
-                        MeshMaterial2d(material),
-                        Transform::from_translation(((center + pos) / 2.0).extend(1.0))
-                            .looking_at((pos - center).extend(0.0), Vec3::Y),
-                        GlobalTransform::default(),
-                        GrabberRope {
-                            player: player_entity,
-                            grabber: nearby.0,
-                        },
-                    ));
+                let sign = if vel.0.dot(tangent) >= 0.0 { 1.0 } else { -1.0 };
+                let mut speed = vel.0.dot(tangent);
+                if speed.abs() < 600.0 {
+                    speed = 600.0 * sign;
                 }
+
+                commands
+                    .entity(player_entity)
+                    .insert(GrabbedBy(nearby.0))
+                    .insert(GrabbedOrbit {
+                        center,
+                        distance,
+                        angle,
+                        initial_speed: speed,
+                    });
+                let mesh = meshes.add(Rectangle::new(4.0, 1.0));
+                let material = materials.add(Color::srgb(0.0, 0.0, 1.0));
+
+                commands.spawn((
+                    Mesh2d(mesh),
+                    MeshMaterial2d(material),
+                    Transform::from_translation(((center + pos) / 2.0).extend(1.0))
+                        .looking_at((pos - center).extend(0.0), Vec3::Y),
+                    GlobalTransform::default(),
+                    GrabberRope {
+                        player: player_entity,
+                        grabber: nearby.0,
+                    },
+                ));
             }
         } else if !is_pressed && grabbed_by.is_some() {
             commands
@@ -272,6 +271,7 @@ fn update_grabbed_players(
     }
 }
 
+#[allow(clippy::disallowed_methods)] // Visual doesn't need determinism
 fn update_grabber_ropes(
     mut commands: Commands,
     rope_query: Query<(Entity, &GrabberRope)>,
@@ -294,10 +294,9 @@ fn update_grabber_ropes(
                     translation: mid.extend(3.0),
                     rotation: rot,
                     scale: Vec3::new(1.0, dir.length(), 1.0),
-                    ..Default::default()
                 });
             }
-            _ => commands.entity(entity).despawn_recursive(),
+            _ => commands.entity(entity).despawn(),
         }
     }
 }
@@ -348,7 +347,7 @@ fn cleanup_grabber_ropes(
 ) {
     for (entity, rope) in rope_query.iter() {
         if let Ok(None) = player_query.get(rope.player) {
-            commands.entity(entity).despawn_recursive();
+            commands.entity(entity).despawn();
         }
     }
 }
